Move out vg circularize
